% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_constructors.R
\name{higher_better}
\alias{higher_better}
\title{Create a higher-is-better scoring rule}
\usage{
higher_better(min_value = NULL, max_value = NULL)
}
\arguments{
\item{min_value}{Optional minimum value for normalization. If NULL,
the minimum will be calculated from the data.}

\item{max_value}{Optional maximum value for normalization. If NULL,
the maximum will be calculated from the data.}
}
\value{
A scoring_rule object of class c("scoring_rule", "higher_better")
}
\description{
Creates a scoring rule object for indicators where higher values indicate
better soil quality. The rule will normalize values using min-max
normalization: (x - min) / (max - min).
}
\examples{
# Create a rule for organic matter (higher is better)
om_rule <- higher_better()

# With custom min/max values
om_rule <- higher_better(min_value = 0, max_value = 5)

}
\seealso{
\code{\link{lower_better}}, \code{\link{optimum_range}},
  \code{\link{threshold_scoring}}
}
