% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherent_adjustment.R
\name{print.inherent_adjustment}
\alias{print.inherent_adjustment}
\title{Print method for inherent_adjustment objects}
\usage{
\method{print}{inherent_adjustment}(x, ...)
}
\arguments{
\item{x}{An \code{inherent_adjustment} object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for inherent_adjustment objects
}
