% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_adequacy.R
\name{print.pca_adequacy}
\alias{print.pca_adequacy}
\title{Print method for pca_adequacy objects}
\usage{
\method{print}{pca_adequacy}(x, ...)
}
\arguments{
\item{x}{A \code{pca_adequacy} object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for pca_adequacy objects
}
