% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_constructors.R
\name{print.scoring_rule}
\alias{print.scoring_rule}
\title{Print method for scoring_rule objects}
\usage{
\method{print}{scoring_rule}(x, ...)
}
\arguments{
\item{x}{A scoring_rule object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for scoring_rule objects
}
