% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularity.R
\name{print.sqi_circularity}
\alias{print.sqi_circularity}
\title{Print method for sqi_circularity objects}
\usage{
\method{print}{sqi_circularity}(x, ...)
}
\arguments{
\item{x}{An \code{sqi_circularity} object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sqi_circularity objects
}
