% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.sqi_stability}
\alias{print.sqi_stability}
\title{Print method for sqi_stability objects}
\usage{
\method{print}{sqi_stability}(x, ...)
}
\arguments{
\item{x}{An \code{sqi_stability} object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sqi_stability objects
}
